Updated Heat Map UI
